# src/main/resources/openapi/notification-service.yaml
openapi: 3.1.0
info:
  title: Notification Service
  description: Shared platform service for email and webhook notifications
  version: 1.0.0
  contact:
    name: Platform Team
servers:
  - url: /api/v1
paths:
...
  /webhook-subscriptions:
    get:
      operationId: listWebhookSubscriptions
      summary: List webhook subscriptions
      tags:
        - webhook-subscriptions
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Paginated list of subscriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createWebhookSubscription
      summary: Register a webhook subscription
      tags:
        - webhook-subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookSubscriptionRequest'
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /webhook-subscriptions/{subscriptionId}:
    get:
      operationId: getWebhookSubscription
      summary: Get subscription details
      tags:
        - webhook-subscriptions
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
      responses:
        '200':
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateWebhookSubscription
      summary: Update a subscription
      tags:
        - webhook-subscriptions
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookSubscriptionRequest'
      responses:
        '200':
          description: Subscription updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscriptionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    delete:
      operationId: deleteWebhookSubscription
      summary: Deactivate a subscription
      tags:
        - webhook-subscriptions
      parameters:
        - $ref: '#/components/parameters/SubscriptionId'
      responses:
        '204':
          description: Subscription deactivated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
...
    SubscriptionId:
      name: subscriptionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
      description: Sort expression (e.g., createdAt,desc)
...
  schemas:
...
    WebhookSubscriptionListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/WebhookSubscriptionResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    CreateWebhookSubscriptionRequest:
      type: object
      required:
        - name
        - callbackUrl
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        callbackUrl:
          type: string
          format: uri
        eventTypes:
          type: array
          items:
            type: string
    UpdateWebhookSubscriptionRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        callbackUrl:
          type: string
          format: uri
        eventTypes:
          type: array
          items:
            type: string
        active:
          type: boolean
    WebhookSubscriptionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        callbackUrl:
          type: string
          format: uri
        eventTypes:
          type: array
          items:
            type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    UnprocessableEntity:
      description: Validation error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...